refactor: extract @otter-agent/rpc package from core
Extract all RPC-related code into a new `@otter-agent/rpc` package,
leaving `@otter-agent/core` as a pure library with no binary or
process-level dependencies.

Moved to @otter-agent/rpc:
- RPC protocol types (RpcCommand, RpcResponse, RpcTransport, etc.)
- RpcHandler (command dispatch, graceful shutdown sequence)
- createRpcSession factory (wires session + handler + UI provider)
- RpcUIProvider (bridges extension UI calls to RPC transport)
- CLI entry point (otter binary, arg parsing, config loading)
- StdioTransport and JSONL serialization
- All associated test files and fixture data

Changes to @otter-agent/core:
- Removed `bin` field and `yaml` dependency from package.json
- Removed all RPC type/value re-exports from index.ts
- Removed RpcUIProvider from ui-providers/index.ts
- Removed UIProvider.rpc() namespace method
- Deleted src/rpc/ and src/cli/ directories

The new package depends on @otter-agent/core,
@otter-agent/environment-registry, and @otter-agent/extension-registry.

Closes #108
