--- packages/otter-agent/src/rpc/index.ts
@@ -1,51 +1,0 @@
-export type {
-	// Commands
-	AbortCommand,
-	CompactCommand,
-	FollowUpCommand,
-	GetCommandsCommand,
-	GetStateCommand,
-	PromptCommand,
-	RpcCommand,
-	RpcCommandType,
-	SetModelCommand,
-	SetThinkingLevelCommand,
-	ShutdownCommand,
-	SteerCommand,
-	// Responses
-	RpcErrorResponse,
-	RpcResponse,
-	RpcSuccessResponse,
-	// Response data
-	RpcCommandInfo,
-	RpcGetCommandsData,
-	RpcResponseDataFor,
-	RpcSessionState,
-	RpcSetModelData,
-	// Events
-	RpcAgentEvent,
-	RpcAgentEventName,
-	RpcExtensionEventName,
-	// Extension UI
-	ExtensionUIConfirmRequest,
-	ExtensionUIDialogRequest,
-	ExtensionUIInputRequest,
-	ExtensionUINotifyRequest,
-	ExtensionUIRequest,
-	ExtensionUIResponse,
-	ExtensionUISelectRequest,
-	// Message unions
-	RpcInboundMessage,
-	RpcOutboundMessage,
-	// Transport
-	RpcTransport,
-} from "./types.js";
-
-export { RpcHandler } from "./rpc-handler.js";
-export type { RpcHandlerOptions } from "./rpc-handler.js";
-
-export { createRpcSession } from "./create-rpc-session.js";
-export type {
-	CreateRpcSessionOptions,
-	CreateRpcSessionResult,
-} from "./create-rpc-session.js";
