+++ packages/rpc/package.json
@@ -1,7 +1,10 @@
 {
-	"name": "@otter-agent/core",
+	"name": "@otter-agent/rpc",
 	"version": "0.0.1",
 	"type": "module",
+	"bin": {
+		"otter": "./dist/cli/cli.js"
+	},
 	"main": "./dist/index.js",
 	"types": "./dist/index.d.ts",
 	"exports": {
@@ -20,13 +23,10 @@
 	"dependencies": {
 		"@mariozechner/pi-agent-core": "^0.64.0",
 		"@mariozechner/pi-ai": "^0.64.0",
+		"@otter-agent/core": "*",
+		"@otter-agent/environment-registry": "*",
+		"@otter-agent/extension-registry": "*",
 		"@sinclair/typebox": "^0.34.0",
-		"better-sqlite3": "^11.0.0",
-		"diff": "^8.0.4",
-		"just-bash": "^2.14.0"
-	},
-	"devDependencies": {
-		"@mariozechner/pi-coding-agent": "^0.64.0",
-		"@types/better-sqlite3": "^7.6.0"
+		"yaml": "^2.7.0"
 	}
 }
