+++ packages/rpc/src/cli/args.test.ts
