+++ packages/rpc/src/cli/auth.test.ts
