+++ packages/rpc/src/cli/cli.ts
