+++ packages/rpc/src/cli/fixtures/bad-config.json
