+++ packages/rpc/src/cli/fixtures/invalid-config.json
