+++ packages/rpc/src/cli/fixtures/missing-path.json
