+++ packages/rpc/src/cli/fixtures/non-existent-path.json
