+++ packages/rpc/src/cli/fixtures/valid-extension-no-config.yaml
