+++ packages/rpc/src/cli/fixtures/valid-extension.yaml
