+++ packages/rpc/src/cli/fixtures/valid-session-manager.json
