+++ packages/rpc/src/cli/jsonl.ts
