+++ packages/rpc/src/cli/load-component.test.ts
@@ -1,7 +1,7 @@
 import { dirname, resolve } from "node:path";
 import { fileURLToPath } from "node:url";
+import { ComponentConfigValidationError } from "@otter-agent/core";
 import { describe, expect, test } from "vitest";
-import { ComponentConfigValidationError } from "../index.js";
 import {
 	ComponentConfigFileError,
 	ComponentLoadError,
