+++ packages/rpc/src/cli/load-extensions.test.ts
