+++ packages/rpc/src/cli/main.ts
@@ -1,5 +1,5 @@
-import type { AgentEnvironment, AuthStorage, SessionManager } from "../index.js";
-import { ModelRegistry } from "../index.js";
+import type { AgentEnvironment, AuthStorage, SessionManager } from "@otter-agent/core";
+import { ModelRegistry } from "@otter-agent/core";
 import { parseCliArgs, printHelp } from "./args.js";
 import { buildAuthStorageFromEnv } from "./auth.js";
 import {
@@ -9,7 +9,7 @@ import {
 	loadComponent,
 } from "./load-component.js";
 import { loadExtensionsFromConfigFiles } from "./load-extensions.js";
-import { runRpcMode } from "./rpc/rpc-mode.js";
+import { runRpcMode } from "./rpc-mode.js";
 
 const VERSION = "0.0.1";
